Split standard-user tasks page into "tasks page" vs "tasks management page"
