+++ src/lib/components/v2/tasks/task_group_utilities.js
@@ -1,4 +1,3 @@
-import { formatMarkdown } from '$lib/common/component_utilities';
 import { getAlertErrorFromResponse } from '$lib/common/errors';
 import compareLoose from 'semver/functions/compare-loose';
 
@@ -156,7 +155,7 @@ function getTaskTableProperties(taskGroup, task) {
 		authors: task.authors,
 		tags: task.tags,
 		input_types: task.input_types,
-		docs_info: formatMarkdown(task.docs_info || '')
+		docs_info: task.docs_info || ''
 	};
 }
